// desktop/src/types.ts
export type TUserInfo = Readonly<{
  id: string
  name: string
  email?: string
  avatar?: string
  provider?: string
}>

export type TAuthStatus = Readonly<{
  authenticated: boolean
  provider?: string
  userInfo?: TUserInfo
}>